arreglados errores en formula para sacar total de bloques, funcionando: DS, IS e ID de mkfile

+++ filesys.h
@@ -29,7 +29,7 @@ private:
     vector<FileData*> file_data_array;
     char *bitmap;
     char *bitmap_inodes;
-    int start_bitmap=0,start_bitmap_inodes=0,start_filetable=0,start_inodes=0,start_datablocks=0;
+    double start_bitmap=0,start_bitmap_inodes=0,start_filetable=0,start_inodes=0,start_datablocks=0;
     vector<QString> disks;
 
     QString root_path = "/";
